% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{DepLabData}
\alias{DepLabData}
\alias{DepLabData-package}
\title{DepLabData- an R package that loads data from the Dephoure Lab at Weill Medical College}
\description{
This package provides 6 datasets. Names of the datasets and a short description are provided below:
WT_trial1: Protein predictions for the first wild type sample
WT_trial2: Protein predictions for the second wild type sample
DN_trial1: Protein predictions for the first D319N transformed sample
DN_trial2: Protein predictions for the second D319N transformed sample
EV_trial1: Protein predictions for the first empty vector sample
EV_trial2: Protein predictions for the second empty vector sample
}
\author{
Julia Wrobel
}
