% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DN_trial1.R
\docType{data}
\name{DN_trial1}
\alias{DN_trial1}
\title{D319N transformed Sample 21}
\description{
Protein predictions for the first D319N transformed sample.
}
\references{
Dephoure, N., and Cavaliere, P.
Weill Cornell Medical College
}
\keyword{datasets}
