% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DN_trial2.R
\docType{data}
\name{DN_trial2}
\alias{DN_trial2}
\title{D319N transformed Sample 2}
\description{
Protein predictions for the second D319N transformed sample.
}
\references{
Dephoure, N., and Cavaliere, P.
Weill Cornell Medical College
}
\keyword{datasets}
