% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV_trial1.R
\docType{data}
\name{EV_trial1}
\alias{EV_trial1}
\title{Empty Vector Sample 1}
\description{
Protein predictions for the first empty vector sample.
}
\references{
Dephoure, N., and Cavaliere, P.
Weill Cornell Medical College
}
\keyword{datasets}
