% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV_trial2.R
\docType{data}
\name{EV_trial2}
\alias{EV_trial2}
\title{Empty Vector Sample 2}
\description{
Protein predictions for the second empty vector sample.
}
\references{
Dephoure, N., and Cavaliere, P.
Weill Cornell Medical College
}
\keyword{datasets}
