% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WT_trial1.R
\docType{data}
\name{WT_trial1}
\alias{WT_trial1}
\title{Wild Type Sample 1}
\description{
Protein predictions for the first wild type sample.
}
\references{
Dephoure, N.,  and Cavaliere, P.
Weill Cornell Medical College
}
\keyword{datasets}
