% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WT_trial2.R
\docType{data}
\name{WT_trial2}
\alias{WT_trial2}
\title{Wild Type Sample 2}
\description{
Protein predictions for the second wild type sample.
}
\references{
Dephoure, N.,  and Cavaliere, P.
Weill Cornell Medical College
}
\keyword{datasets}
